modified:   __pycache__/logic.cpython-38.pyc 	modified:   cli.py 	modified:   logic.py
Additional features added:

The game should be playable by a single player or 2 players.
In the case of a single player, the player should play against a 'bot'. For implementing this bot, you can just choose feasible random numbers as the inputs to the game from the bot (you can build more sophisticated logic for the inputs from the bots if you are interested).
In the case of two players, the players should take turns entering the inputs in the CLI.
The functionality should be implemented through objected-oriented dependency injection.
The above steps are repeated until a winner emerges.